refactor(users): change providers folder strucutre
