+++ src/modules/users/providers/hash/hash-provider.interface.ts
