+++ src/modules/users/providers/jwt/jwt-provider.interface.ts
