+++ src/modules/users/users.module.ts
@@ -5,14 +5,14 @@ import { Model } from 'mongoose';
 
 import { User, UserSchema } from './entities/users.entity';
 import { UsersFactory } from './entities/users.factory';
-import { IHashProvider } from './providers/hash-provider.interface';
-import { BcryptHashProvider } from './providers/implementations/bcrypt-hash.provider';
-import { HandleBarsTemplateEngineProvider } from './providers/implementations/handlebars-template-engine.provider';
-import { JsonWebTokenProvider } from './providers/implementations/jsonwebtoken.provider';
-import { NodeMailerMailProvider } from './providers/implementations/nodemailer-mail.provider';
-import { IJwtProvider } from './providers/jwt-provider.interface';
-import { IMailProvider } from './providers/mail-provider.interface';
-import { ITemplateEngineProvider } from './providers/template-engine-provider.interface';
+import { BcryptHashProvider } from './providers/hash/bcrypt/bcrypt-hash.provider';
+import { IHashProvider } from './providers/hash/hash-provider.interface';
+import { JsonWebTokenProvider } from './providers/jwt/jsonwebtoken/jsonwebtoken.provider';
+import { IJwtProvider } from './providers/jwt/jwt-provider.interface';
+import { IMailProvider } from './providers/mail/mail-provider.interface';
+import { NodeMailerMailProvider } from './providers/mail/nodemailer/nodemailer-mail.provider';
+import { HandleBarsTemplateEngineProvider } from './providers/template-engine/handlebars/handlebars-template-engine.provider';
+import { ITemplateEngineProvider } from './providers/template-engine/template-engine-provider.interface';
 import { UsersMongoRepository } from './repositories/mongo/users-mongo.repository';
 import { IUsersRepository } from './repositories/users-repository.interface';
 import { SignupUserUseCase } from './usecases';
